[REFACTOR: IDEA draft and post validation schema]

+++ src/app/modules/comments/comments.service.ts
@@ -1,21 +1,24 @@
-import { Comment } from '@prisma/client';
+import { Comment, Role } from '@prisma/client';
 import prisma from '../../config/prisma';
 import AppError from '../../errors/AppError';
 import { httpStatus } from '../../utils/httpStatus';
 import { JwtPayload } from 'jsonwebtoken';
 
-const createCommentsIntoDB = async (payload: Partial<Comment>,user:JwtPayload) => {
-  if ( !payload.ideaId || !payload.content) {
+const createCommentsIntoDB = async (
+  payload: Partial<Comment>,
+  user: JwtPayload
+) => {
+  if (!payload.ideaId || !payload.content) {
     throw new AppError(httpStatus.BAD_REQUEST, 'Required fields missing');
   }
- const filterData={
-  content: payload.content,
-  ideaId: payload.ideaId,
-  userId: user.id,
-  parentId: payload.parentId || null,
-}
+  const filterData = {
+    content: payload.content,
+    ideaId: payload.ideaId,
+    userId: user.id,
+    parentId: payload.parentId || null,
+  };
   const result = await prisma.comment.create({
-    data:filterData ,
+    data: filterData,
     include: {
       user: { select: { name: true } },
     },
@@ -55,13 +58,31 @@ const getCommentByIdeaIdFromDB = async (ideaId: string) => {
   return comments;
 };
 
+const deleteCommentFromDB = async (id: string, authUser: JwtPayload) => {
+  const userId = authUser.id;
 
-const deleteCommentFromDB = async (id: string,user:JwtPayload) => {
-  const userId=user.id;
-  const comment = await prisma.comment.delete({
-    where: { id ,userId},
+  const comment = await prisma.comment.findUnique({
+    where: { id },
   });
-  return comment;
+
+  if (!comment) {
+    throw new AppError(httpStatus.NOT_FOUND, 'Comment not found');
+  }
+
+  if (authUser.role !== Role.ADMIN) {
+    if (comment.userId !== userId) {
+      throw new AppError(
+        httpStatus.NOT_ACCEPTABLE,
+        "Not premited to delete other's comments!"
+      );
+    }
+  }
+
+  const deletedComment = await prisma.comment.delete({
+    where: { id },
+  });
+
+  return deletedComment;
 };
 
 export const commentService = {
